adding scenatio and refactoring code
